New file structure #41
