+++ Config/DefaultHttpGPT.ini
@@ -4,3 +4,32 @@
 +FunctionRedirects=(OldName="HttpGPTRequest.SendMessages", NewName="HttpGPTRequest.SendMessages_CustomOptions")
 +FunctionRedirects=(OldName="HttpGPTSettings.GetDefaultSettings", NewName="HttpGPTSettings.GetDefaultOptions")
 +FunctionRedirects=(OldName="HttpGPTSettings.SetDefaultSettings", NewName="HttpGPTSettings.SetDefaultOptions")
+
+
+;v1.5.0
+; Classes
++ClassRedirects=(OldName="/Script/HttpGPT.HttpGPTRequest", NewName="/Script/HttpGPTChatModule.HttpGPTChatRequest")
++ClassRedirects=(OldName="/Script/HttpGPT.HttpGPTHelper", NewName="/Script/HttpGPTCommonModule.HttpGPTHelper")
++ClassRedirects=(OldName="/Script/HttpGPT.HttpGPTSettings", NewName="/Script/HttpGPTCommonModule.HttpGPTSettings")
+
+; Delegates
++ClassRedirects=(OldName="/Script/HttpGPT.HttpGPTResponseDelegate", NewName="/Script/HttpGPTChatModule.HttpGPTChatResponseDelegate")
++ClassRedirects=(OldName="/Script/HttpGPT.HttpGPTGenericDelegate", NewName="/Script/HttpGPTCommonModule.HttpGPTGenericDelegate")
+
+; Enumerations
++EnumRedirects=(OldName="/Script/HttpGPT.HttpGPTRole",NewName="/Script/HttpGPTCommonModule.HttpGPTChatRole")
++EnumRedirects=(OldName="/Script/HttpGPT.HttpGPTModel",NewName="/Script/HttpGPTCommonModule.HttpGPTChatModel")
+
+; Structures
++StructRedirects=(OldName="/Script/HttpGPT.HttpGPTResponse",NewName="/Script/HttpGPTCommonModule.HttpGPTChatResponse")
++StructRedirects=(OldName="/Script/HttpGPT.HttpGPTMessage",NewName="/Script/HttpGPTCommonModule.HttpGPTChatMessage")
++StructRedirects=(OldName="/Script/HttpGPT.HttpGPTChoice",NewName="/Script/HttpGPTCommonModule.HttpGPTChatChoice")
++StructRedirects=(OldName="/Script/HttpGPT.HttpGPTUsage",NewName="/Script/HttpGPTCommonModule.HttpGPTChatUsage")
++StructRedirects=(OldName="/Script/HttpGPT.HttpGPTError",NewName="/Script/HttpGPTCommonModule.HttpGPTChatError")
++StructRedirects=(OldName="/Script/HttpGPT.HttpGPTOptions",NewName="/Script/HttpGPTCommonModule.HttpGPTChatOptions")
+
+; Properties
++PropertyRedirects=(OldName="/Script/HttpGPTCommonModule.HttpGPTChatOptions.APIKey", NewName="/Script/HttpGPTCommonModule.HHttpGPTSettings.APIKey")
+
+; Functions
++FunctionRedirects=(OldName="/Script/HttpGPTCommonModule.HttpGPTHelper.CastToHttpGPTRequest", NewName="/Script/HttpGPTChatModule.HttpGPTChatHelper.CastToHTTPGPTChatRequest")
