+++ HttpGPT.uplugin
@@ -1,10 +1,10 @@
 {
 	"FileVersion": 3,
-	"Version": 12,
-	"VersionName": "1.4.7",
-	"FriendlyName": "HttpGPT - ChatGPT integrated in the Engine",
-	"Description": "HttpGPT is an Unreal Engine plugin that facilitates integration with Chat GPT through asynchronous REST requests, making it easy for developers to communicate with the chatbot. HttpGPT also includes a new Editor Tool to integrate Chat GPT directly in the Engine.",
-	"Category": "Messaging",
+	"Version": 13,
+	"VersionName": "1.5.0",
+	"FriendlyName": "HttpGPT - GPT Integration",
+	"Description": "HttpGPT is an Unreal Engine plugin that facilitates integration with OpenAI's GPT based services (ChatGPT and DALL-E) through asynchronous REST requests, making it easy for developers to communicate with the chatbot. HttpGPT also includes Editor Tools to integrate Chat GPT and DALL-E image generation directly in the Engine.",
+	"Category": "Game Features",
 	"CreatedBy": "Lucas Vilas-Boas",
 	"CreatedByURL": "https://github.com/lucoiso",
 	"DocsURL": "https://github.com/lucoiso/UEHttpGPT/wiki",
@@ -16,7 +16,7 @@
 	"Installed": false,
 	"Modules": [
 		{
-			"Name": "HttpGPT",
+			"Name": "HttpGPTChatModule",
 			"Type": "Runtime",
 			"LoadingPhase": "Default",
 			"WhitelistPlatforms": [
@@ -29,7 +29,33 @@
 			]
 		},
 		{
-			"Name": "HttpGPTEditor",
+			"Name": "HttpGPTImageModule",
+			"Type": "Runtime",
+			"LoadingPhase": "Default",
+			"WhitelistPlatforms": [
+				"Win64",
+				"Mac",
+				"Linux",
+				"IOS",
+				"Android",
+				"HoloLens"
+			]
+		},
+		{
+			"Name": "HttpGPTCommonModule",
+			"Type": "Runtime",
+			"LoadingPhase": "Default",
+			"WhitelistPlatforms": [
+				"Win64",
+				"Mac",
+				"Linux",
+				"IOS",
+				"Android",
+				"HoloLens"
+			]
+		},
+		{
+			"Name": "HttpGPTEditorModule",
 			"Type": "Editor",
 			"WhitelistPlatforms": [
 				"Win64",
