--- Source/HttpGPT/Private/HttpGPT.cpp
@@ -1,19 +1,0 @@
-// Author: Lucas Vilas-Boas
-// Year: 2023
-// Repo: https://github.com/lucoiso/UEHttpGPT
-
-#include "HttpGPT.h"
-
-#define LOCTEXT_NAMESPACE "FHttpGPTModule"
-
-void FHttpGPTModule::StartupModule()
-{
-}
-
-void FHttpGPTModule::ShutdownModule()
-{
-}
-
-#undef LOCTEXT_NAMESPACE
-	
-IMPLEMENT_MODULE(FHttpGPTModule, HttpGPT)
