--- Source/HttpGPT/Public/LogHttpGPT.h
@@ -1,14 +1,0 @@
-// Author: Lucas Vilas-Boas
-// Year: 2023
-// Repo: https://github.com/lucoiso/UEHttpGPT
-
-#pragma once
-
-#include <Logging/LogMacros.h>
-
-/**
- *
- */
-
-DECLARE_LOG_CATEGORY_EXTERN(LogHttpGPT, Display, All);
-DECLARE_LOG_CATEGORY_EXTERN(LogHttpGPT_Internal, NoLogging, All);
