+++ Source/HttpGPTChatModule/Private/HttpGPTChatModule.cpp
@@ -1,0 +1,19 @@
+// Author: Lucas Vilas-Boas
+// Year: 2023
+// Repo: https://github.com/lucoiso/UEHttpGPT
+
+#include "HttpGPTChatModule.h"
+
+#define LOCTEXT_NAMESPACE "FHttpGPTChatModule"
+
+void FHttpGPTChatModule::StartupModule()
+{
+}
+
+void FHttpGPTChatModule::ShutdownModule()
+{
+}
+
+#undef LOCTEXT_NAMESPACE
+	
+IMPLEMENT_MODULE(FHttpGPTChatModule, HttpGPTChatModule)
