+++ Source/HttpGPTChatModule/Public/HttpGPTChatModule.h
@@ -1,0 +1,15 @@
+// Author: Lucas Vilas-Boas
+// Year: 2023
+// Repo: https://github.com/lucoiso/UEHttpGPT
+
+#pragma once
+
+#include <CoreMinimal.h>
+#include <Modules/ModuleManager.h>
+
+class FHttpGPTChatModule : public IModuleInterface
+{
+public:
+	virtual void StartupModule() override;
+	virtual void ShutdownModule() override;
+};
