+++ Source/HttpGPTCommonModule/Private/HttpGPTCommonModule.cpp
@@ -1,0 +1,19 @@
+// Author: Lucas Vilas-Boas
+// Year: 2023
+// Repo: https://github.com/lucoiso/UEHttpGPT
+
+#include "HttpGPTCommonModule.h"
+
+#define LOCTEXT_NAMESPACE "FHttpGPTCommonModule"
+
+void FHttpGPTCommonModule::StartupModule()
+{
+}
+
+void FHttpGPTCommonModule::ShutdownModule()
+{
+}
+
+#undef LOCTEXT_NAMESPACE
+	
+IMPLEMENT_MODULE(FHttpGPTCommonModule, HttpGPTCommonModule)
