+++ Source/HttpGPTCommonModule/Private/Structures/HttpGPTImageTypes.cpp
@@ -1,0 +1,5 @@
+// Author: Lucas Vilas-Boas
+// Year: 2023
+// Repo: https://github.com/lucoiso/UEHttpGPT
+
+#include "Structures/HttpGPTImageTypes.h"
