+++ Source/HttpGPTCommonModule/Public/LogHttpGPT.h
@@ -1,0 +1,14 @@
+// Author: Lucas Vilas-Boas
+// Year: 2023
+// Repo: https://github.com/lucoiso/UEHttpGPT
+
+#pragma once
+
+#include <Logging/LogMacros.h>
+
+/**
+ *
+ */
+
+HTTPGPTCOMMONMODULE_API DECLARE_LOG_CATEGORY_EXTERN(LogHttpGPT, Display, All);
+HTTPGPTCOMMONMODULE_API DECLARE_LOG_CATEGORY_EXTERN(LogHttpGPT_Internal, NoLogging, All);
