+++ Source/HttpGPTCommonModule/Public/Structures/HttpGPTImageTypes.h
@@ -1,0 +1,7 @@
+// Author: Lucas Vilas-Boas
+// Year: 2023
+// Repo: https://github.com/lucoiso/UEHttpGPT
+
+// These structures are defined in the common module due to being used in both modules, to avoid cyclic dependencies.
+
+#pragma once
