+++ Source/HttpGPTImageModule/Private/HttpGPTImageModule.cpp
@@ -1,0 +1,19 @@
+// Author: Lucas Vilas-Boas
+// Year: 2023
+// Repo: https://github.com/lucoiso/UEHttpGPT
+
+#include "HttpGPTImageModule.h"
+
+#define LOCTEXT_NAMESPACE "FHttpGPTImageModule"
+
+void FHttpGPTImageModule::StartupModule()
+{
+}
+
+void FHttpGPTImageModule::ShutdownModule()
+{
+}
+
+#undef LOCTEXT_NAMESPACE
+	
+IMPLEMENT_MODULE(FHttpGPTImageModule, HttpGPTImageModule)
