+++ Source/HttpGPTImageModule/Private/Tasks/HttpGPTImageRequest.cpp
@@ -1,0 +1,5 @@
+// Author: Lucas Vilas-Boas
+// Year: 2023
+// Repo: https://github.com/lucoiso/UEHttpGPT
+
+#include "Tasks/HttpGPTImageRequest.h"
